# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Wiki REST API
  version: "1.0.0"
  description: >
    Wiki システム用 REST API。
    ページ管理、アセット管理、ページロック（TTL/トークン）を提供する。
servers:
  - url: /api
security:
  - basicAuth: []
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  parameters:
    PageId:
      name: page_id
      in: path
      required: true
      schema: { type: string }
...
    LockAuthHeader:
      name: X-Lock-Authentication
      in: header
      required: true
      schema: { type: string }
      description: 'token={lock_token}'
  schemas:
    ErrorResponse:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          description: 人間可読なエラー理由
    IdResponse:
      type: object
      required: [id]
      properties:
        id: { type: string }
...
paths:
  /pages:
    post:
      summary: ページの作成
      parameters:
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        "201":
          description: 作成成功
          headers:
            Location: { schema: { type: string } }
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IdResponse" }
        "400":
          description: パス不正
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: 既存ページあり
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /pages/{page_id}:
    delete:
      summary: ページ削除
      parameters:
        - $ref: "#/components/parameters/PageId"
        - $ref: "#/components/parameters/LockAuthHeader"
      responses:
        "204":
          description: 削除成功
        "403":
          description: トークン不正
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: ページ不存在
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "410":
          description: 既に削除済み
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "423":
          description: ロック中（トークン未指定）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...